+++ backend/src/handler/create_game.rs
@@ -1,8 +1,8 @@
 use crate::gamestate::game::Game;
-use crate::jwt::generate_jwt;
 use crate::repo::game_repo::GameRepo;
 use crate::repo::address_repo::AddressRepo;
-use crate::InMemoryGameRepo;
+use crate::repo::authorization_repo::AuthorizationRepo;
+use crate::{InMemoryGameRepo};
 use actix_web::{post, web, HttpResponse, Responder};
 use serde::Deserialize;
 use serde::Serialize;
@@ -23,6 +23,7 @@ pub struct GameCreateResponse {
 #[post("/game")]
 pub async fn create_game(
     game_repo: web::Data<Arc<Mutex<InMemoryGameRepo>>>,
+    authorization_repo: web::Data<Arc<AuthorizationRepo>>,
     address_repo: web::Data<Arc<AddressRepo>>,
     body: web::Json<GameCreateData>,
 ) -> impl Responder {
@@ -35,9 +36,9 @@ pub async fn create_game(
 
     let game = Game::new(author_name);
     let game_id = game.id.clone();
-    let jwt = generate_jwt(author_name, &game_id);
 
     game_repo.lock().unwrap().add_game(game);
+    let jwt = authorization_repo.generate_jwt(author_name, &game_id);
 
     HttpResponse::Created().json(GameCreateResponse {
         gameID: game_id,
