+++ backend/src/repo/mod.rs
@@ -1,2 +1,3 @@
 pub mod game_repo;
 pub mod address_repo;
+pub mod authorization_repo;
